refactor(services): service creation for global-desc
creation of global-desc.service.ts to fetch description data and display
it to the UI
