+++ catagram_frontend/src/app/pages/merise/merise.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import diagramsData from '../../data/diagrams.json'
+import diagramsData from '../../../assets/data/diagrams.json'
 @Component({
   selector: 'app-merise',
   templateUrl: './merise.component.html',
