+++ catagram_frontend/src/app/pages/uml/uml.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import diagramsData from '../../data/diagrams.json'
+import diagramsData from '../../../assets/data/diagrams.json'
 @Component({
   selector: 'app-uml',
   templateUrl: './uml.component.html',
