+++ catagram_frontend/src/assets/data/diagrams.json
