+++ catagram_frontend/src/assets/data/global-desc.json
